+++ database/migrations/2023_05_18_122941_create_production_orders_products_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
 			$table->foreignIdFor(\App\Models\ProductionOrder::class, 'production_order_id');
 			$table->foreignIdFor(\App\Models\Product::class, 'product_id');
-			$table->float('quantity_needed');
-			$table->float('quantity_transferred');
+			$table->double('quantity_needed');
+			$table->double('quantity_transferred');
 	        $table->enum('status', array_keys(config('requested.warehouse_orders.status')));
             $table->timestamps();
         });
